example_service/requirements.txt
# FastAPI Building Blocks
fastapi-building-blocks>=0.1.0

# Server
uvicorn[standard]>=0.30.0

# Settings management
pydantic-settings>=2.0.0

# Email validation
email-validator>=2.0.0

# Optional dependencies
python-multipart>=0.0.9

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
